restApi/requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-model-utils==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
